+++ src/Vlog.jsx
@@ -1,0 +1,22 @@
+import React from 'react'
+
+const Vlog = () => {
+  return (
+    <>
+      <h2 style={{ marginTop: "40px", fontFamily: "cursive" }}>Vlog</h2>
+      <div className="video">
+        <iframe
+          width="560"
+          height="315"
+          src="https://www.youtube.com/embed/B9krT5Lq-1U?list=PLBn82aS9YRQJ3Apw0cLYCBfC-vOCU5BkG"
+          title="YouTube video player"
+          frameBorder="0"
+          allow="accelerometer; clipboard-write; encrypted-media; gyroscope; picture-in-picture"
+          allowFullScreen
+        ></iframe>
+      </div>
+    </>
+  )
+}
+
+export default Vlog
